UPDATE: Refactor code by moving common types to a separate file in `src/types/common.ts`.
