+++ src/utils/common.ts
@@ -1,4 +1,4 @@
-import { ColsType, RowsType } from '@/components/TableForm/TableForm';
+import { RowsType } from '@/types/common';
 
 export const toClassName = (array: Array<string | number | boolean>) => array.filter(Boolean).join(' ');
 
@@ -118,9 +118,6 @@ export const italicRegex = (string: string) => {
   return newString;
 };
 
-export type GetCurrentRowsReturnType = {
-  getCurrentRows: () => RowsType;
-};
 export const getCurrentRows = () => {
   const table = document.querySelector('table');
   const trs = table!.querySelectorAll('tr');
@@ -136,9 +133,6 @@ export const getCurrentRows = () => {
   return rows;
 };
 
-export type GetCurrentColsReturnType = {
-  getCurrentCols: () => ColsType;
-};
 export const getCurrentCols = () => {
   const table = document.querySelector('table');
   const ths = Array.from(table!.querySelectorAll('th')).map((th) => getInputValue(th));
